float -> double, logger fixed

+++ src/ColorScheme/BaseColorAdder.java
@@ -3,25 +3,25 @@ package ColorScheme;
 public class BaseColorAdder implements ColorAdder {
 
     @Override
-    public float boundColor(float original, float add_color) {
-        float color = original + add_color;
-        if (color < 0.0f) {
+    public double boundColor(double original, double add_color) {
+        double color = original + add_color;
+        if (color < 0.0) {
             add_color = -original;
         }
-        else if (color > 1.0f) {
+        else if (color > 1.0) {
             add_color = 1 - original;
         }
         return add_color;
     }
 
     @Override
-    public float addColor(float original, float add_color) {
-        float color = original + add_color;
-        if (color < 0.0f) {
+    public double addColor(double original, double add_color) {
+        double color = original + add_color;
+        if (color < 0.0) {
             color = 0;
         }
-        else if (color > 1.0f) {
-            color = 1.0f;
+        else if (color > 1.0) {
+            color = 1.0;
         }
         return color;
     }
