+++ src/Estimators/BlackLossEstimator.java
@@ -4,11 +4,11 @@ import java.util.ArrayList;
 
 public class BlackLossEstimator extends PixelLossEstimator {
     @Override
-    public float calc_loss(float original, float canvas) {
+    public double calc_loss(double original, double canvas) {
         return Math.max(0, original - canvas);
     }
 
-    public BlackLossEstimator(float[][] image, ArrayList<float[][]> features) {
+    public BlackLossEstimator(double[][] image, ArrayList<double[][]> features) {
         super(image, features);
     }
 }
