+++ src/Generators/BasicLineDrawer.java
@@ -15,7 +15,7 @@ public class BasicLineDrawer {
         this.colorAdder = colorAdder;
     }
 
-    public void applyToCanvas(float[][] canvas) {
+    public void applyToCanvas(double[][] canvas) {
         for (GrayPoint p : this.points) {
             canvas[p.y][p.x] = p.color;
         }
@@ -25,25 +25,25 @@ public class BasicLineDrawer {
         return this.points;
     }
 
-    public void updateLossEstimator(float[][] canvas, LossEstimator lossEstimator) {
+    public void updateLossEstimator(double[][] canvas, LossEstimator lossEstimator) {
         lossEstimator.update_loss(canvas, this.points);
     }
 
-    private float[] buildLine(float x1, float y1, float x2, float y2) {
-        float a = y1 - y2;
-        float b = x2 - x1;
-        float c = x1 * y2 - x2 * y1;
-        return new float[] {a, b, c};
+    private double[] buildLine(double x1, double y1, double x2, double y2) {
+        double a = y1 - y2;
+        double b = x2 - x1;
+        double c = x1 * y2 - x2 * y1;
+        return new double[] {a, b, c};
     }
 
-    private boolean isOnLine(int x, int y, float x1, float y1, float x2, float y2) {
-        float[] line = buildLine(x1, y1, x2, y2);
-        float[] vertex_x = new float[] {x, x, x + 1, x + 1};
-        float[] vertex_y = new float[] {y, y + 1, y + 1, y};
+    private boolean isOnLine(int x, int y, double x1, double y1, double x2, double y2) {
+        double[] line = buildLine(x1, y1, x2, y2);
+        double[] vertex_x = new double[] {x, x, x + 1, x + 1};
+        double[] vertex_y = new double[] {y, y + 1, y + 1, y};
         for (int i = 0; i < vertex_x.length; ++i) {
             for (int j = i + 1; j < vertex_x.length; ++j) {
-                float res1 = line[0] * vertex_x[i] + line[1] * vertex_y[i] + line[2];
-                float res2 = line[0] * vertex_x[j] + line[1] * vertex_y[j] + line[2];
+                double res1 = line[0] * vertex_x[i] + line[1] * vertex_y[i] + line[2];
+                double res2 = line[0] * vertex_x[j] + line[1] * vertex_y[j] + line[2];
                 if (res1 < 0 && res2 > 0 || res1 > 0 && res2 < 0) {
                     return true;
                 }
@@ -52,11 +52,11 @@ public class BasicLineDrawer {
         return false;
     }
 
-    public void fabricateSegment(float[][] canvas, float x1, float y1, float x2, float y2, float color) {
+    public void fabricateSegment(double[][] canvas, double x1, double y1, double x2, double y2, double color) {
         this.points.clear();
         if (x1 > x2) {
-            float tmp_x = x1;
-            float tmp_y = y1;
+            double tmp_x = x1;
+            double tmp_y = y1;
             x1 = x2;
             y1 = y2;
             x2 = tmp_x;
@@ -80,7 +80,7 @@ public class BasicLineDrawer {
             try_points_y = new int[] {0, -1, -1};
         }
 
-        float col = this.colorAdder.addColor(canvas[cur_y][cur_x], color);
+        double col = this.colorAdder.addColor(canvas[cur_y][cur_x], color);
         this.points.add(new GrayPoint(cur_x, cur_y, col));
 
         while (cur_x != end_x || cur_y != end_y) {
@@ -104,7 +104,7 @@ public class BasicLineDrawer {
         }
     }
 
-    public void drawSegment(float[][] canvas, float x1, float y1, float x2, float y2, float color, LossEstimator lossEstimator) {
+    public void drawSegment(double[][] canvas, double x1, double y1, double x2, double y2, double color, LossEstimator lossEstimator) {
         fabricateSegment(canvas, x1, y1, x2, y2, color);
         applyToCanvas(canvas);
         updateLossEstimator(canvas, lossEstimator);
