+++ src/ImageProcessing/ImageTransformer.java
@@ -2,19 +2,19 @@ package ImageProcessing;
 
 public class ImageTransformer {
 
-    private static float gaussFunction(int x, int y, float sigma) {
+    private static double gaussFunction(int x, int y, double sigma) {
         double pow_denom = (double)sigma;
         pow_denom = 2 * pow_denom * pow_denom;
         double pow = -(double)(x*x + y*y) / pow_denom;
-        return (float)(Math.exp(pow) / (2 * Math.PI * sigma));
+        return (double)(Math.exp(pow) / (2 * Math.PI * sigma));
     }
 
-    private static float[][] buildGaussKernel(int kernel_size, float sigma) {
+    private static double[][] buildGaussKernel(int kernel_size, double sigma) {
         if (kernel_size % 2 == 0) {
             kernel_size++;
         }
         int center = kernel_size / 2;
-        float[][] kernel = new float[kernel_size][kernel_size];
+        double[][] kernel = new double[kernel_size][kernel_size];
         for (int i = 0; i < kernel_size; ++i) {
             for (int j = 0; j < kernel_size; ++j) {
                 kernel[i][j] = gaussFunction(i - center, j - center, sigma);
@@ -23,8 +23,8 @@ public class ImageTransformer {
         return kernel;
     }
 
-    private static float[][] buildSobelKernel(int axis) {
-        float[][] kernel = new float[3][3];
+    private static double[][] buildSobelKernel(int axis) {
+        double[][] kernel = new double[3][3];
         if (axis == 0) {
             kernel[0][0] = -1;
             kernel[0][1] = 0;
@@ -50,8 +50,8 @@ public class ImageTransformer {
         return kernel;
     }
 
-    private static float[][] buildScharrKernel(int axis) {
-        float[][] kernel = new float[3][3];
+    private static double[][] buildScharrKernel(int axis) {
+        double[][] kernel = new double[3][3];
         if (axis == 0) {
             kernel[0][0] = -3;
             kernel[0][1] = 0;
@@ -77,10 +77,10 @@ public class ImageTransformer {
         return kernel;
     }
 
-    private static float applyKernelToCell(float[][] image, float[][] kernel, int x, int y) {
-        float res = 0;
-        float kernel_sum = 0;
-        float missed_sum = 0;
+    private static double applyKernelToCell(double[][] image, double[][] kernel, int x, int y) {
+        double res = 0;
+        double kernel_sum = 0;
+        double missed_sum = 0;
         int half_kernel_size = kernel.length / 2;
         for (int i = 0; i < kernel.length; ++i) {
             for (int j = 0; j < kernel[0].length; ++j) {
@@ -101,8 +101,8 @@ public class ImageTransformer {
         return res * kernel_sum / (kernel_sum - missed_sum);
     }
 
-    private static float[][] applyKernel(float[][] image, float[][] kernel) {
-        float[][] result = new float[image.length][image[0].length];
+    private static double[][] applyKernel(double[][] image, double[][] kernel) {
+        double[][] result = new double[image.length][image[0].length];
         for (int y = 0; y < result.length; ++y) {
             for (int x = 0; x < result[0].length; ++x) {
                 result[y][x] = applyKernelToCell(image, kernel, x, y);
@@ -111,18 +111,18 @@ public class ImageTransformer {
         return result;
     }
 
-    public static float[][] gaussBlur(float[][] image, int kernel_size, float sigma) {
-        float[][] kernel = buildGaussKernel(kernel_size, sigma);
+    public static double[][] gaussBlur(double[][] image, int kernel_size, double sigma) {
+        double[][] kernel = buildGaussKernel(kernel_size, sigma);
         return applyKernel(image, kernel);
     }
 
-    public static float[][] sobel(float[][] image, int axis) {
-        float[][] kernel = buildSobelKernel(axis);
+    public static double[][] sobel(double[][] image, int axis) {
+        double[][] kernel = buildSobelKernel(axis);
         return applyKernel(image, kernel);
     }
 
-    public static float[][] scharr(float[][] image, int axis) {
-        float[][] kernel = buildScharrKernel(axis);
+    public static double[][] scharr(double[][] image, int axis) {
+        double[][] kernel = buildScharrKernel(axis);
         return applyKernel(image, kernel);
     }
 }
